float to double

+++ Move.java
@@ -2,10 +2,10 @@ public class Move{
 
     public String _name, _type;
     public int _basePow;
-    public float _acc;
+    public double _acc;
     public Boolean _isPhysical, _makesContact;
 
-    public Move(String name, String type, int basePow, float acc, Boolean physical, Boolean contact){
+    public Move(String name, String type, int basePow, double acc, Boolean physical, Boolean contact){
 	_name = name;
 	_type = type;
 	_basePow = basePow;
